+++ controllers/project.controller.js
@@ -1,13 +1,35 @@
 const db = require("../models");
 const Project = db.projects;
+const Offer = db.offers;
+const Order = db.orders;
+const Product = db.products;
+const ProjectFile = db.projectFiles;
 const Op = db.Sequelize.Op;
 
 // Create and Save a new Project
 exports.create = (req, res) => {
   // Validate request
-  if (!req.body.title) {
+  if (!req.body.name) {
     res.status(400).send({
-      message: "Content can not be empty!"
+      message: "name can not be empty!"
+    });
+    return;
+  }
+  if (!req.body.description) {
+    res.status(400).send({
+      message: "description can not be empty!"
+    });
+    return;
+  }
+  if (!req.body.createdBy) {
+    res.status(400).send({
+      message: "createdBy can not be empty!"
+    });
+    return;
+  }
+  if (!req.body.customerId) {
+    res.status(400).send({
+      message: "customerId can not be empty!"
     });
     return;
   }
@@ -22,7 +44,8 @@ exports.create = (req, res) => {
     isCompleted: req.body.isCompleted ? req.body.isCompleted : false,
     hasReceipt: req.body.hasReceipt ? req.body.hasReceipt : false,
     hasInvoice: req.body.hasInvoice ? req.body.hasInvoice : false,
-    createdBy: req.body.createdBy ?? null,
+    createdBy: req.body.createdBy,
+    customerId: req.body.customerId,
     updatedBy: null
   };
 
@@ -74,6 +97,12 @@ exports.findOne = (req, res) => {
 
 // Update a Project by the id in the request
 exports.update = (req, res) => {
+  if (!req.body.updatedBy) {
+    res.status(400).send({
+      message: "updatedBy can not be empty!"
+    });
+    return;
+  }
   const id = req.params.id;
 
   Project.update(req.body, {
@@ -99,6 +128,12 @@ exports.update = (req, res) => {
 
 // Delete a Project with the specified id in the request
 exports.delete = (req, res) => {
+  if (!req.body.updatedBy) {
+    res.status(400).send({
+      message: "updatedBy can not be empty!"
+    });
+    return;
+  }
   const id = req.params.id;
 
   Project.destroy({
@@ -118,6 +153,59 @@ exports.delete = (req, res) => {
     .catch(err => {
       res.status(500).send({
         message: "Could not delete Project with id=" + id
+      });
+    });
+};
+// Extras
+exports.getProjectFiles = (req, res) => {
+  const id = req.params.id;
+  var condition =  { projectId: id  };
+  ProjectFile.findAll({ where: condition })
+    .then(data => {
+        res.send(data);
+    })
+    .catch(err => {
+      res.status(500).send({
+        message: "Error retrieving Project ProjectFiless with id=" + id
+      });
+    });
+};
+exports.getProjectOffers = (req, res) => {
+  const id = req.params.id;
+  var condition =  { projectId: id  };
+  Offer.findAll({ where: condition })
+    .then(data => {
+        res.send(data);
+    })
+    .catch(err => {
+      res.status(500).send({
+        message: "Error retrieving Project Offers with id=" + id
+      });
+    });
+};
+exports.getProjectOrders = (req, res) => {
+  const id = req.params.id;
+  var condition =  { projectId: id  };
+  Order.findAll({ where: condition })
+    .then(data => {
+        res.send(data);
+    })
+    .catch(err => {
+      res.status(500).send({
+        message: "Error retrieving Project Orders with id=" + id
+      });
+    });
+};
+exports.getProjectProducts = (req, res) => {
+  const id = req.params.id;
+  var condition =  { projectId: id  };
+  Product.findAll({ where: condition })
+    .then(data => {
+        res.send(data);
+    })
+    .catch(err => {
+      res.status(500).send({
+        message: "Error retrieving Project Products with id=" + id
       });
     });
 };
