+++ controllers/semiProductDesignFile.controller.js
@@ -5,9 +5,21 @@ const Op = db.Sequelize.Op;
 // Create and Save a new SemiProductDesignFile
 exports.create = (req, res) => {
   // Validate request
-  if (!req.body.title) {
+  if (!req.body.name) {
     res.status(400).send({
-      message: "Content can not be empty!"
+      message: "name can not be empty!"
+    });
+    return;
+  }
+  if (!req.body.createdBy) {
+    res.status(400).send({
+      message: "createdBy can not be empty!"
+    });
+    return;
+  }
+  if (!req.body.semiProductId) {
+    res.status(400).send({
+      message: "semiProductId can not be empty!"
     });
     return;
   }
@@ -15,7 +27,8 @@ exports.create = (req, res) => {
   // Create a SemiProductDesignFile
   const semiProductDesignFile = {
     name: req.body.name,
-    createdBy: req.body.createdBy ?? null,
+    createdBy: req.body.createdBy,
+    semiProductId: req.body.semiProductId,
     updatedBy: null
   };
 
@@ -67,6 +80,18 @@ exports.findOne = (req, res) => {
 
 // Update a SemiProductDesignFile by the id in the request
 exports.update = (req, res) => {
+  if (!req.body.updatedBy) {
+    res.status(400).send({
+      message: "updatedBy can not be empty!"
+    });
+    return;
+  }
+  if (!req.body.updatedBy) {
+    res.status(400).send({
+      message: "updatedBy can not be empty!"
+    });
+    return;
+  }
   const id = req.params.id;
 
   SemiProductDesignFile.update(req.body, {
@@ -92,6 +117,18 @@ exports.update = (req, res) => {
 
 // Delete a SemiProductDesignFile with the specified id in the request
 exports.delete = (req, res) => {
+  if (!req.body.updatedBy) {
+    res.status(400).send({
+      message: "updatedBy can not be empty!"
+    });
+    return;
+  }
+  if (!req.body.updatedBy) {
+    res.status(400).send({
+      message: "updatedBy can not be empty!"
+    });
+    return;
+  }
   const id = req.params.id;
 
   SemiProductDesignFile.destroy({
